Reimplementation of the PNG texture cache in FastLoader
